lot of useless rewriting , and i written some unit tests
