+++ .gitignore
@@ -1,10 +1,4 @@
-CCCP/lolcat*
-CCCP/bin/*
-CCCP/build/*
-CCCP/sources/*
 .git
 .vs
 .vscode
-CCCP/run.sh
-CCCP/DATA_DIR
-CCCP/PKG_DIR
+CCCP/tests/env/*
