+++ CCCP/include/deps.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int check_dependencies (const std::vector<std::string>& dependencies, const std::string& DATA_DIR);
