+++ CCCP/include/install.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void install_source(const std::string &PName,const std::string &PKG_DIR,const std::string &DATA_FILE,const std::string &DATA_DIR,const std::string &MAKE_DIR,const std::string &BUILD_DIR,const std::string &LOG_DIR,const std::string &ROOT);
