+++ CCCP/include/move.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void move_binaries( std::vector<std::string> files ,std::string ROOT);
