+++ CCCP/src/deps.cpp
@@ -1,7 +1,7 @@
 #include <string>
 #include <vector>
-
-#include "../lib/log.hpp"
+#include <iostream> 
+#include <unistd.h>
 
 
 // This function will check if all dependencies of a package are installed
@@ -19,7 +19,8 @@ int check_dependencies (const std::vector<std::string>& dependencies, const std:
         for (int i = 0; i < dependencies.size(); i++)
         {
             // TODO: implement vesrion checking
-            if (std::filesystem::exists(DATA_DIR + dependencies[i] + ".spm"))
+            // is use the C function because my clangd server doesnt recognize the fucking std::filesystem thing and it bothered me 
+            if (access((DATA_DIR + dependencies[i] + ".spm").c_str(),F_OK))
             {
                 std::cout << "dependency " << dependencies[i] << " is installed.\n";
             }
@@ -27,7 +28,7 @@ int check_dependencies (const std::vector<std::string>& dependencies, const std:
                 std::cout << "dependency " << dependencies[i] << " is not installed.\n";
                 //install automatically the needed dependency
                 //To do this run install_package
-                install_package(dependencies[i]);
+                //install_package(dependencies[i]);
             }
         }
     }
