+++ CCCP/tests/test.cpp
@@ -1,0 +1,18 @@
+#include <iostream>
+#include <string>
+
+#include "../include/locations.h"
+
+std::string tests();
+
+int main()
+{
+    std::cout << "Starting CCCP testing \n";
+
+    std::cout << tests() << std::endl;
+    return 0;
+}
+std::string tests()
+{
+    
+}
